+++ zarrs_storage/src/storage_adapter/async_to_sync.rs
@@ -58,12 +58,12 @@ impl<TStorage: ?Sized, TBlockOn: AsyncToSyncBlockOn> AsyncToSyncStorageAdapter<T
 impl<TStorage: ?Sized + AsyncReadableStorageTraits, TBlockOn: AsyncToSyncBlockOn>
     ReadableStorageTraits for AsyncToSyncStorageAdapter<TStorage, TBlockOn>
 {
-    fn get_byte_ranges<'a>(
+    fn get_partial_many<'a>(
         &'a self,
         key: &StoreKey,
         byte_ranges: ByteRangeIterator<'a>,
     ) -> Result<MaybeBytesIterator<'a>, StorageError> {
-        let results = self.block_on(self.storage.get_byte_ranges(key, byte_ranges))?;
+        let results = self.block_on(self.storage.get_partial_many(key, byte_ranges))?;
         if let Some(results) = results {
             let results = self.block_on(results.collect::<Vec<_>>());
             Ok(Some(Box::new(results.into_iter())))
@@ -116,8 +116,8 @@ impl<TStorage: ?Sized + AsyncWritableStorageTraits, TBlockOn: AsyncToSyncBlockOn
         self.block_on(self.storage.erase(key))
     }
 
-    fn erase_values(&self, keys: &[StoreKey]) -> Result<(), StorageError> {
-        self.block_on(self.storage.erase_values(keys))
+    fn erase_many(&self, keys: &[StoreKey]) -> Result<(), StorageError> {
+        self.block_on(self.storage.erase_many(keys))
     }
 
     fn erase_prefix(&self, prefix: &StorePrefix) -> Result<(), StorageError> {