+++ zarrs_storage/src/store/memory_store.rs
@@ -63,7 +63,7 @@ impl ReadableStorageTraits for MemoryStore {
         }
     }
 
-    fn get_byte_ranges<'a>(
+    fn get_partial_many<'a>(
         &'a self,
         key: &StoreKey,
         byte_ranges: ByteRangeIterator<'a>,